requirements.txt
streamlit>=1.33.0
python-dotenv>=1.0.1
google-generativeai>=0.7.0s